Restructure coverage as separate command
