+++ libs/natlint/src/main/java/org/amshove/natlint/NatLint.java
@@ -1,12 +1,26 @@
 package org.amshove.natlint;
 
 import org.amshove.natlint.cli.AnalyzeCommand;
+import org.amshove.natlint.cli.CoverageCommand;
 import picocli.CommandLine;
+import picocli.CommandLine.UnmatchedArgumentException;
 
 public class NatLint
 {
 	public static void main(String[] args)
 	{
-		System.exit(new CommandLine(new AnalyzeCommand()).execute(args));
+		CommandLine coverageCommandLine = new CommandLine(new CoverageCommand());
+		CommandLine analyzeCommandLine = new CommandLine(new AnalyzeCommand());
+
+		try {
+			coverageCommandLine.parseArgs(args);
+			System.exit(coverageCommandLine.execute(args));
+		}
+		catch (UnmatchedArgumentException _)
+		{
+			// This is normal;
+		}
+
+		System.exit(analyzeCommandLine.execute(args));
 	}
 }
