+++ libs/natlint/src/main/java/org/amshove/natlint/cli/CliAnalyzer.java
@@ -13,7 +13,6 @@ import org.amshove.natlint.linter.NaturalLinter;
 import org.amshove.natparse.IDiagnostic;
 import org.amshove.natparse.ReadOnlyList;
 import org.amshove.natparse.infrastructure.ActualFilesystem;
-import org.amshove.natparse.lexing.Lexer;
 import org.amshove.natparse.lexing.TokenList;
 import org.amshove.natparse.natural.INaturalModule;
 import org.amshove.natparse.natural.project.*;
@@ -28,301 +27,96 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class CliAnalyzer
+public class CliAnalyzer extends ProjectAnalyzer
 {
-	private final ActualFilesystem filesystem;
 	private final IDiagnosticSink diagnosticSink;
 	private final boolean disableLinting;
-	private Path workingDirectory;
 	private final AnalyzerPredicates predicates;
-	private final FileStatusSink fileStatusSink;
-	private final AnalyzerOutputFlags outputFlags;
-	private Map<String, AtomicInteger> totalDiagnosticsById = new HashMap<>();
-
-	public CliAnalyzer(Path workingDirectory, IDiagnosticSink sink, FileStatusSink fileStatusSink, AnalyzerPredicates predicates, boolean disableLinting, AnalyzerOutputFlags outputFlags)
-	{
-		this.workingDirectory = workingDirectory;
-		this.predicates = predicates;
-		filesystem = new ActualFilesystem();
-		diagnosticSink = sink;
-		this.fileStatusSink = fileStatusSink;
-		this.disableLinting = disableLinting;
-		this.outputFlags = outputFlags;
-	}
-
-	public int run()
-	{
-		var filesystem = new ActualFilesystem();
-
-		while (!workingDirectory.getRoot().equals(workingDirectory) && filesystem.findNaturalProjectFile(workingDirectory).isEmpty())
-		{
-			workingDirectory = workingDirectory.getParent();
-		}
-
-		var projectFile = filesystem.findNaturalProjectFile(workingDirectory);
-
-		if (workingDirectory.getRoot().equals(workingDirectory) || projectFile.isEmpty())
-		{
-			throw new RuntimeException("Project root could not be determined. .natural or _naturalBuild file not found");
-		}
-
-		var editorconfigPath = projectFile.get().getParent().resolve(".editorconfig");
-		if (editorconfigPath.toFile().exists())
-		{
-			LinterContext.INSTANCE.updateEditorConfig(new EditorConfigParser().parse(filesystem.readFile(editorconfigPath)));
-		}
-
-		System.out.printf(
-			"""
-			     .@@@@@@@@@@@@@@@&
-			    /@@@@@@@@@@@@@@@@@.          %s
-			     @@@@@@@@*@@@@@@@@           Version: %s
-			  ....###############            Project file: %s
-			......###.@/##.@.####......      %s
-			     .###############
-			     .###############
-			     .###############
-			       ############.
-			           ....
-			%n""",
-			CliAnalyzer.class.getPackage().getImplementationTitle(),
-			CliAnalyzer.class.getPackage().getImplementationVersion(),
-			projectFile.get().getFileName(),
-			editorconfigPath.toFile().exists() ? ".editorconfig picked up" : ""
-		);
-
-		predicates.printSettings();
-
-		return analyze(projectFile.get());
-	}
-
-	private SlowestModule slowestLexedModule = new SlowestModule(Long.MIN_VALUE, "NONE");
-	private SlowestModule slowestParsedModule = new SlowestModule(Long.MIN_VALUE, "NONE");
 	private SlowestModule slowestLintedModule = new SlowestModule(Long.MIN_VALUE, "NONE");
-	private final AtomicInteger filesChecked = new AtomicInteger();
-	private final AtomicInteger totalDiagnostics = new AtomicInteger();
-	private final AtomicInteger exceptions = new AtomicInteger();
-	private final AtomicLong linesOfCode = new AtomicLong();
-	private long maxMemoryInBytes = 0L;
 
-	private int analyze(Path projectFilePath)
+	public CliAnalyzer(Path workingDirectory, IDiagnosticSink sink, FileStatusSink fileStatusSink, AnalyzerPredicates predicates, boolean disableLinting, AnalyzerOutputFlags outputFlags, String sourceEncoding)
 	{
-		var indexStartTime = System.currentTimeMillis();
-		var project = new BuildFileProjectReader(filesystem).getNaturalProject(projectFilePath);
-		new NaturalProjectFileIndexer().indexProject(project);
-		var indexEndTime = System.currentTimeMillis();
-
-		var startCheck = System.currentTimeMillis();
-		for (var library : project.getLibraries())
-		{
-			library.files().parallelStream().forEach(file ->
-			{
-				if (file.getFiletype() == NaturalFileType.DDM)
-				{
-					// DdmParser isn't called in the CLI. DDMs will be parsed on demand.
-					return;
-				}
-
-				if (file.isFailedOnInit())
-				{
-					fileStatusSink.printError(file.getPath(), MessageType.INDEX_EXCEPTION, file.getInitException());
-					return;
-				}
-
-				if (!predicates.shouldAnalyzeFile(file))
-				{
-					fileStatusSink.printStatus(file.getPath(), MessageType.FILE_EXCLUDED);
-					return;
-				}
-
-				filesChecked.incrementAndGet();
-				var allDiagnosticsInFile = new ArrayList<IDiagnostic>();
-
-				var tokens = lex(file, allDiagnosticsInFile);
-				if (tokens == null)
-				{
-					return;
-				}
-
-				var module = parse(file, tokens, allDiagnosticsInFile);
-				if (module == null)
-				{
-					diagnosticSink.printDiagnostics(filesChecked.get(), file.getPath(), allDiagnosticsInFile);
-					return;
-				}
-
-				if (tokens.sourceHeader().getProgrammingMode() == NaturalProgrammingMode.REPORTING)
-				{
-					fileStatusSink.printStatus(file.getPath(), MessageType.REPORTING_TYPE);
-					diagnosticSink.printDiagnostics(filesChecked.get(), file.getPath(), allDiagnosticsInFile);
-					return;
-				}
-
-				if (!disableLinting && module.programmingMode() != NaturalProgrammingMode.REPORTING)
-				{
-					var linterDiagnostics = lint(file, module, allDiagnosticsInFile);
-					if (linterDiagnostics == null)
-					{
-						return;
-					}
-				}
-
-				var totalDiagnosticsInFileById = allDiagnosticsInFile.stream()
-					.collect(Collectors.groupingBy(IDiagnostic::id, Collectors.counting()));
-				for (var diagnosticId : totalDiagnosticsInFileById.keySet())
-				{
-					if (!totalDiagnosticsById.containsKey(diagnosticId))
-					{
-						totalDiagnosticsById.put(diagnosticId, new AtomicInteger());
-					}
-					totalDiagnosticsById.get(diagnosticId)
-						.addAndGet(totalDiagnosticsInFileById.get(diagnosticId).intValue());
-				}
-
-				totalDiagnostics.addAndGet(allDiagnosticsInFile.size());
-				diagnosticSink.printDiagnostics(filesChecked.get(), file.getPath(), allDiagnosticsInFile);
-				fileStatusSink.printStatus(file.getPath(), MessageType.SUCCESS);
-			});
-			var currentMemory = Runtime.getRuntime().totalMemory();
-			if (currentMemory > maxMemoryInBytes)
-			{
-				maxMemoryInBytes = currentMemory;
-			}
-		}
-
-		var endCheck = System.currentTimeMillis();
-
-		var missingStartTime = System.currentTimeMillis();
-		registerMissingFiles(project);
-		var missingEndTime = System.currentTimeMillis();
-
-		var indexTime = indexEndTime - indexStartTime;
-		var checkTime = endCheck - startCheck;
-		var missTime = missingEndTime - missingStartTime;
-		var totalTime = indexTime + checkTime + missTime;
-
-		var totalTimeSeconds = totalTime / 1000;
-		System.out.println();
-		System.out.println("Done.");
-		System.out.printf("Index time: %d ms%n", indexTime);
-		System.out.printf("Check time: %d ms%n", checkTime);
-		System.out.printf("Miss time : %d ms%n", missTime);
-		System.out.printf("Total: %d ms (%ds)%n", totalTime, totalTimeSeconds);
-		System.out.println();
-		System.out.printf("Files checked: %,d%n", filesChecked.get());
-		System.out.printf("Lines of code: %,d%n", linesOfCode.get());
-		System.out.printf("LoC/s: %,d%n", totalTimeSeconds > 0 ? (linesOfCode.get() / totalTimeSeconds) : linesOfCode.get());
-		System.out.println();
-		System.out.printf("Total diagnostics: %,d%n", totalDiagnostics.get());
-		System.out.println("Exceptions: " + exceptions.get());
-		System.out.println();
-		System.out.println("Slowest lexed module: " + slowestLexedModule);
-		System.out.println("Slowest parsed module: " + slowestParsedModule);
-		System.out.println("Slowest linted module: " + (disableLinting ? "disabled" : slowestLintedModule));
-		System.out.println();
-		System.out.printf("Peak memory usage: %.2f Mib%n", maxMemoryInBytes / 1024.0 / 1024.0);
-		var gcs = 0L;
-		var gcTime = 0L;
-		for (var bean : ManagementFactory.getGarbageCollectorMXBeans())
-		{
-			gcs += bean.getCollectionCount();
-			if (bean.getCollectionTime() > 0)
-			{
-				gcTime += bean.getCollectionTime();
-			}
-		}
-		System.out.printf("Number of GCs: %d%n", gcs);
-		System.out.printf("GC time: %ds%n", (gcTime / 1000));
-		System.out.println();
-		if (outputFlags.showDiagnosticStats())
-		{
-			System.out.println("Total diagnostics by ID");
-			for (var diagnosticId : totalDiagnosticsById.keySet().stream().sorted().toList())
-			{
-				System.out.printf("%s: %,d%n", diagnosticId, totalDiagnosticsById.get(diagnosticId).get());
-			}
-			System.out.println();
-		}
-
-		return totalDiagnostics.get() > 0 ? 1 : 0;
+		super(workingDirectory, fileStatusSink, outputFlags);
+		this.predicates = predicates;
+		diagnosticSink = sink;
+		this.disableLinting = disableLinting;
 	}
 
-	private List<? extends IDiagnostic> filterDiagnostics(ReadOnlyList<? extends IDiagnostic> diagnostics)
+	@Override
+	protected void printUnderBanner()
+	{
+		predicates.printSettings();
+	}
+
+	@Override
+	protected boolean shouldAnalyzeFile(NaturalFile file)
+	{
+		if (file.getFiletype() == NaturalFileType.DDM)
+		{
+			// DdmParser isn't called in the CLI. DDMs will be parsed on demand.
+			return false;
+		}
+
+		if (file.isFailedOnInit())
+		{
+			fileStatusSink.printError(file.getPath(), MessageType.INDEX_EXCEPTION, file.getInitException());
+			return false;
+		}
+
+		if (!predicates.shouldAnalyzeFile(file))
+		{
+			fileStatusSink.printStatus(file.getPath(), MessageType.FILE_EXCLUDED);
+			return false;
+		}
+
+		return true;
+	}
+
+	@Override
+	protected void analyzeFile(NaturalFile file, ArrayList<IDiagnostic> allDiagnosticsInFile, TokenList tokens, INaturalModule module)
+	{
+		if (module == null)
+		{
+			diagnosticSink.printDiagnostics(filesChecked.get(), file.getPath(), allDiagnosticsInFile);
+			return;
+		}
+
+		if (tokens.sourceHeader().getProgrammingMode() == NaturalProgrammingMode.REPORTING)
+		{
+			fileStatusSink.printStatus(file.getPath(), MessageType.REPORTING_TYPE);
+			diagnosticSink.printDiagnostics(filesChecked.get(), file.getPath(), allDiagnosticsInFile);
+			return;
+		}
+
+		if (!disableLinting && module.programmingMode() != NaturalProgrammingMode.REPORTING)
+		{
+			var linterDiagnostics = lint(file, module, allDiagnosticsInFile);
+			if (linterDiagnostics == null)
+			{
+				return;
+			}
+		}
+
+		var totalDiagnosticsInFileById = allDiagnosticsInFile.stream()
+			.collect(Collectors.groupingBy(IDiagnostic::id, Collectors.counting()));
+		for (var diagnosticId : totalDiagnosticsInFileById.keySet())
+		{
+			if (!totalDiagnosticsById.containsKey(diagnosticId))
+			{
+				totalDiagnosticsById.computeIfAbsent(diagnosticId, _ -> new AtomicInteger())
+					.addAndGet(totalDiagnosticsInFileById.get(diagnosticId).intValue());
+			}
+		}
+
+		totalDiagnostics.addAndGet(allDiagnosticsInFile.size());
+		diagnosticSink.printDiagnostics(filesChecked.get(), file.getPath(), allDiagnosticsInFile);
+		fileStatusSink.printStatus(file.getPath(), MessageType.SUCCESS);
+	}
+
+	@Override
+	protected List<IDiagnostic> filterDiagnostics(ReadOnlyList<IDiagnostic> diagnostics)
 	{
 		return diagnostics.stream().filter(predicates::shouldPrintDiagnostic).toList();
-	}
-
-	private TokenList lex(NaturalFile file, ArrayList<IDiagnostic> allDiagnosticsInFile)
-	{
-		try
-		{
-			var lexer = new Lexer();
-			var lexStart = System.currentTimeMillis();
-			var tokens = lexer.lex(filesystem.readFile(file.getPath()), file.getPath());
-			var lexEnd = System.currentTimeMillis();
-			countLinesOfCode(tokens);
-			if (slowestLexedModule.milliseconds < lexEnd - lexStart)
-			{
-				slowestLexedModule = new SlowestModule(lexEnd - lexStart, file.getProjectRelativePath().toString());
-			}
-
-			var diagnostics = filterDiagnostics(tokens.diagnostics());
-			fileStatusSink.printDiagnostics(file.getPath(), MessageType.LEX_FAILED, diagnostics);
-			allDiagnosticsInFile.addAll(diagnostics);
-			return tokens;
-		}
-		catch (Exception e)
-		{
-			fileStatusSink.printError(file.getPath(), MessageType.LEX_EXCEPTION, e);
-			exceptions.incrementAndGet();
-			System.out.println(file.getPath());
-			e.printStackTrace();
-			return null;
-		}
-	}
-
-	private void countLinesOfCode(TokenList tokens)
-	{
-		var previousLine = -1;
-		var totalLines = 0;
-		for (var token : tokens)
-		{
-			if (token.line() != previousLine)
-			{
-				totalLines++;
-				previousLine = token.line();
-			}
-		}
-
-		linesOfCode.addAndGet(totalLines);
-	}
-
-	private INaturalModule parse(NaturalFile file, TokenList tokens, ArrayList<IDiagnostic> allDiagnosticsInFile)
-	{
-		try
-		{
-			var parser = new NaturalParser();
-			var parseStart = System.currentTimeMillis();
-			var module = parser.parse(file, tokens);
-			var parseEnd = System.currentTimeMillis();
-			if (slowestParsedModule.milliseconds < parseEnd - parseStart)
-			{
-				slowestParsedModule = new SlowestModule(parseEnd - parseStart, file.getProjectRelativePath().toString());
-			}
-
-			var diagnostics = filterDiagnostics(module.diagnostics());
-			fileStatusSink.printDiagnostics(file.getPath(), MessageType.PARSE_FAILED, diagnostics);
-			allDiagnosticsInFile.addAll(diagnostics);
-			return module;
-		}
-		catch (Exception e)
-		{
-			fileStatusSink.printError(file.getPath(), MessageType.PARSE_EXCEPTION, e);
-			exceptions.incrementAndGet();
-			System.out.println(file.getPath());
-			e.printStackTrace();
-			return null;
-		}
 	}
 
 	private ReadOnlyList<LinterDiagnostic> lint(NaturalFile file, INaturalModule module, ArrayList<IDiagnostic> allDiagnosticsInFile)
@@ -333,7 +127,7 @@ public class CliAnalyzer
 			var lintStart = System.currentTimeMillis();
 			var linterDiagnostics = linter.lint(module);
 			var lintEnd = System.currentTimeMillis();
-			if (slowestLintedModule.milliseconds < lintEnd - lintStart)
+			if (slowestLintedModule.milliseconds() < lintEnd - lintStart)
 			{
 				slowestLintedModule = new SlowestModule(lintEnd - lintStart, file.getProjectRelativePath().toString());
 			}
@@ -353,37 +147,4 @@ public class CliAnalyzer
 		}
 	}
 
-	private void registerMissingFiles(NaturalProject project)
-	{
-		if (!fileStatusSink.isEnabled())
-		{
-			return;
-		}
-
-		System.err.println("Started registration of missing files");
-		var root = project.getRootPath().resolve("Natural-Libraries");
-		System.out.println("Root: " + root.toString());
-
-		try (var stream = Files.walk(root))
-		{
-			stream
-				.filter(path -> !Files.isDirectory(path))
-				.forEach(path -> fileStatusSink.printStatus(path, MessageType.FILE_MISSING));
-			System.err.println("Finished registration of missing files");
-		}
-		catch (Exception e)
-		{
-			System.err.println("Registration of missing files failed");
-			e.printStackTrace();
-		}
-	}
-
-	record SlowestModule(long milliseconds, String module)
-	{
-		@Override
-		public String toString()
-		{
-			return "%dms (%s)".formatted(milliseconds, module);
-		}
-	}
 }
